+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Devtools now work
 - Tab name updates correctly for internal pages
 - Add shortcut to reopen last closed tab
+- Auto focus address bar
 
 ## v0.7.0 - 2026-02-15